# docs/07-data-api-events/specs/openapi.platform.yaml
openapi: 3.1.2
info:
  title: Korean Equity RWA PoC Platform API
  version: 1.0.0
  description: 합성 데이터만 사용하는 7단계 플랫폼 HTTP 계약이다. 상태변경 명령의 202 응답은 업무 완료가 아니다.
servers:
  - url: http://localhost:3000/api/v1
security:
  - syntheticBearer: []
...
paths:
...
  /positions:
    get:
      tags: [CommonQueries]
      operationId: listPositions
      summary: 고객 또는 기관 범위의 권리와 토큰 조회
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 고객 또는 역할 범위의 권리, 토큰과 지급청구 투영
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedPositions' }
  /activities:
    get:
      tags: [CommonQueries]
      operationId: listActivities
      summary: 주문과 권리 및 자금 활동내역 조회
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 역할 범위의 활동내역
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedActivities' }
  /workflows/{workflowId}:
    get:
      tags: [CommonQueries]
      operationId: getWorkflow
      summary: 업무의 독립 상태축 조회
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: 독립 상태축과 투영 완전성
          content:
            application/json:
              schema:
                $ref: './schemas/domain.schema.json#/$defs/WorkflowView'
        '404': { $ref: '#/components/responses/Error' }
  /workflows/{workflowId}/timeline:
    get:
      tags: [CommonQueries]
      operationId: getWorkflowTimeline
      summary: 업무 증거와 정정 흐름 조회
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: 명령, 기관사실, 승인, 체인결과와 정정 계보
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [workflowId, items, projection]
                properties:
                  workflowId: { type: string, format: uuid }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/TimelineItem' }
                  projection:
                    $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata'
...
  /market-maker/hedges/{hedgeId}/decisions:
    post:
      tags: [InstitutionCommands]
      operationId: decideMarketMakerHedge
      summary: 시장조성자 헤지 제출 여부 결정
      parameters:
        - name: hedgeId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './schemas/domain.schema.json#/$defs/InstitutionDecision' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    syntheticBearer:
      type: http
      scheme: bearer
      bearerFormat: Synthetic-JWT
    adapterSignature:
      type: apiKey
      in: header
      name: X-Adapter-Key-Id
      description: 본문 안의 Ed25519 signature와 함께 검증한다.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1, maxLength: 512 }
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Accepted:
      description: 비동기 업무 접수
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/AcceptedCommand' }
    Error:
      description: 안정적인 기계 오류와 한국어 다음 행동
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/ErrorResponse' }
...
  schemas:
...
    Position:
      type: object
      additionalProperties: false
      required: [securityId, settledRights, pendingRights, lockedRights, burnPendingTokens, projection]
      properties:
        securityId: { $ref: './schemas/common.schema.json#/$defs/SecurityId' }
        settledRights: { $ref: './schemas/common.schema.json#/$defs/NonNegativeIntegerString' }
        pendingRights: { $ref: './schemas/common.schema.json#/$defs/NonNegativeIntegerString' }
        lockedRights: { $ref: './schemas/common.schema.json#/$defs/NonNegativeIntegerString' }
        burnPendingTokens: { $ref: './schemas/common.schema.json#/$defs/NonNegativeIntegerString' }
        cashClaim: { $ref: './schemas/common.schema.json#/$defs/Money' }
        projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
    PagedPositions:
      type: object
      additionalProperties: false
      required: [items, projection]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Position' } }
        nextCursor: { type: string }
        projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
    TimelineItem:
      type: object
      additionalProperties: false
      required: [eventId, eventType, occurredAt, labelKo]
      properties:
        eventId: { type: string, format: uuid }
        eventType: { type: string }
        occurredAt: { type: string, format: date-time }
        labelKo: { type: string, minLength: 1 }
        correctsEventId: { type: string, format: uuid }
    PagedActivities:
      type: object
      additionalProperties: false
      required: [items, projection]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TimelineItem' } }
        nextCursor: { type: string }
        projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
...